requirements.txt
langchain-groq
langchain-core
langchain-openai
langchain
sentence-transformers
faiss-cpu
numpy
python-dotenv
